requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
black==22.1.0
cffi==1.15.0
click==8.0.4
dnspython==2.1.0
fastapi==0.70.0
fastapi-csrf-protect==0.2.1
flake8==4.0.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.1.1
mccabe==0.6.1
motor==2.5.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.4.0
PyJWT==2.3.0
pymongo==3.12.3
python-decouple==3.3
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tomli==2.0.1
typing_extensions==4.1.1
uvicorn==0.14.0
